refactor(plugins) integrates handler files
Removes smaller code files by integrating them into the `handler` files.
Reducing overall file clutter
