--- kong/plugins/file-log/log.lua
@@ -1,65 +1,0 @@
--- Copyright (C) Mashape, Inc.
-
-local ffi = require "ffi"
-local cjson = require "cjson"
-local fd_util = require "kong.plugins.file-log.fd_util"
-local system_constants = require "lua_system_constants"
-local basic_serializer = require "kong.plugins.log-serializers.basic"
-
-local ngx_timer = ngx.timer.at
-local string_len = string.len
-local O_CREAT = system_constants.O_CREAT()
-local O_WRONLY = system_constants.O_WRONLY()
-local O_APPEND = system_constants.O_APPEND()
-local S_IRUSR = system_constants.S_IRUSR()
-local S_IWUSR = system_constants.S_IWUSR()
-local S_IRGRP = system_constants.S_IRGRP()
-local S_IROTH = system_constants.S_IROTH()
-
-local oflags = bit.bor(O_WRONLY, O_CREAT, O_APPEND)
-local mode = bit.bor(S_IRUSR, S_IWUSR, S_IRGRP, S_IROTH)
-
-ffi.cdef[[
-int open(char * filename, int flags, int mode);
-int write(int fd, void * ptr, int numbytes);
-
-char *strerror(int errnum);
-]]
-
-local function string_to_char(str)
-  return ffi.cast("uint8_t*", str)
-end
-
--- Log to a file
--- @param `premature`
--- @param `conf`     Configuration table, holds http endpoint details
--- @param `message`  Message to be logged
-local function log(premature, conf, message)
-  message = cjson.encode(message).."\n"
-
-  local fd = fd_util.get_fd(conf.path)
-  if not fd then
-    fd = ffi.C.open(string_to_char(conf.path), oflags, mode)
-    if fd < 0 then
-      local errno = ffi.errno()
-      ngx.log(ngx.ERR, "[file-log] failed to open the file: ", ffi.string(ffi.C.strerror(errno)))
-    else
-      fd_util.set_fd(conf.path, fd)
-    end
-  end
-
-  ffi.C.write(fd, string_to_char(message), string_len(message))
-end
-
-local _M = {}
-
-function _M.execute(conf)
-  local message = basic_serializer.serialize(ngx)
-
-  local ok, err = ngx_timer(0, log, conf, message)
-  if not ok then
-    ngx.log(ngx.ERR, "[file-log] failed to create timer: ", err)
-  end
-end
-
-return _M
