+++ kong/plugins/ip-restriction/handler.lua
@@ -1,8 +1,10 @@
 local BasePlugin = require "kong.plugins.base_plugin"
-local init_worker = require "kong.plugins.ip-restriction.init_worker"
-local access = require "kong.plugins.ip-restriction.access"
+local responses = require "kong.tools.responses"
+local iputils = require "resty.iputils"
 
 local IpRestrictionHandler = BasePlugin:extend()
+
+IpRestrictionHandler.PRIORITY = 990
 
 function IpRestrictionHandler:new()
   IpRestrictionHandler.super.new(self, "ip-restriction")
@@ -10,14 +12,28 @@ end
 
 function IpRestrictionHandler:init_worker()
   IpRestrictionHandler.super.init_worker(self)
-  init_worker.execute()
+  local ok, err = iputils.enable_lrucache()
+  if not ok then
+    ngx.log(ngx.ERR, "[ip-restriction] Could not enable lrucache: ", err)
+  end
 end
 
 function IpRestrictionHandler:access(conf)
   IpRestrictionHandler.super.access(self)
-  access.execute(conf)
-end
+  local block = false
+  local remote_addr = ngx.var.remote_addr
 
-IpRestrictionHandler.PRIORITY = 990
+  if conf._blacklist_cache and #conf._blacklist_cache > 0 then
+    block = iputils.ip_in_cidrs(remote_addr, conf._blacklist_cache)
+  end
+
+  if conf._whitelist_cache and #conf._whitelist_cache > 0 then
+    block = not iputils.ip_in_cidrs(remote_addr, conf._whitelist_cache)
+  end
+
+  if block then
+    return responses.send_HTTP_FORBIDDEN("Your IP address is not allowed")
+  end
+end
 
 return IpRestrictionHandler
