+++ kong/plugins/udp-log/handler.lua
@@ -1,8 +1,36 @@
-local log = require "kong.plugins.udp-log.log"
+local cjson = require "cjson"
 local BasePlugin = require "kong.plugins.base_plugin"
 local basic_serializer = require "kong.plugins.log-serializers.basic"
 
 local UdpLogHandler = BasePlugin:extend()
+
+UdpLogHandler.PRIORITY = 1
+
+local function log(premature, conf, message)
+  local host = conf.host
+  local port = conf.port
+  local timeout = conf.timeout
+
+  local sock = ngx.socket.udp()
+  sock:settimeout(timeout)
+
+  local ok, err = sock:setpeername(host, port)
+  if not ok then
+    ngx.log(ngx.ERR, "failed to connect to "..host..":"..tostring(port)..": ", err)
+    return
+  end
+
+  ok, err = sock:send(cjson.encode(message))
+  if not ok then
+    ngx.log(ngx.ERR, "failed to send data to ".. host..":"..tostring(port)..": ", err)
+  end
+
+  ok, err = sock:close()
+  if not ok then
+    ngx.log(ngx.ERR, "failed to close connection from "..host..":"..tostring(port)..": ", err)
+    return
+  end
+end
 
 function UdpLogHandler:new()
   UdpLogHandler.super.new(self, "udp-log")
@@ -12,9 +40,10 @@ function UdpLogHandler:log(conf)
   UdpLogHandler.super.log(self)
 
   local message = basic_serializer.serialize(ngx)
-  log.execute(conf, message)
+  local ok, err = ngx.timer.at(0, log, conf, message)
+  if not ok then
+    ngx.log(ngx.ERR, "failed to create timer: ", err)
+  end
 end
-
-UdpLogHandler.PRIORITY = 1
 
 return UdpLogHandler
